Fix test using consumer to work after rebase
