+++ core/src/test/scala/integration/kafka/api/ConsumerRebootstrapTest.scala
@@ -16,17 +16,22 @@
  */
 package kafka.api
 
+import kafka.api.ConsumerRebootstrapTest._
 import kafka.server.QuorumTestHarness.getTestQuorumAndGroupProtocolParametersClassicGroupProtocolOnly_ZK_implicit
 import kafka.utils.{TestInfoUtils, TestUtils}
+import org.apache.kafka.clients.CommonClientConfigs
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerRecord}
-import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions.{assertEquals, assertThrows}
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.{Arguments, MethodSource}
 
+import java.time.Duration
 import java.util.{Collections, stream}
+import java.util.concurrent.TimeUnit
+import java.util.concurrent.TimeoutException
 
 class ConsumerRebootstrapTest extends RebootstrapTest {
-  @ParameterizedTest(name = s"${TestInfoUtils.TestWithParameterizedQuorumAndGroupProtocolNames}.useRebootstrapTriggerMs={2}")
+  @ParameterizedTest(name = RebootstrapTestName)
   @MethodSource(Array("rebootstrapTestParams"))
   def testRebootstrap(quorum: String, groupProtocol: String, useRebootstrapTriggerMs: Boolean): Unit = {
     sendRecords(10, 0)
@@ -79,6 +84,40 @@ class ConsumerRebootstrapTest extends RebootstrapTest {
     consumeAndVerifyRecords(consumer, 10, 20, startingKeyAndValueIndex = 20, startingTimestamp = 20)
   }
 
+  @ParameterizedTest(name = RebootstrapTestName)
+  @MethodSource(Array("rebootstrapTestParams"))
+  def testRebootstrapDisabled(quorum: String, groupProtocol: String, useRebootstrapTriggerMs: Boolean): Unit = {
+    server1.shutdown()
+    server1.awaitShutdown()
+
+    val configOverrides = clientOverrides(useRebootstrapTriggerMs)
+    configOverrides.put(CommonClientConfigs.METADATA_RECOVERY_STRATEGY_CONFIG, "none")
+    if (useRebootstrapTriggerMs)
+      configOverrides.put(CommonClientConfigs.METADATA_RECOVERY_REBOOTSTRAP_TRIGGER_MS_CONFIG, "1000")
+
+    val producer = createProducer(configOverrides = configOverrides)
+    val consumer = createConsumer(configOverrides = configOverrides)
+    val adminClient = createAdminClient(configOverrides = configOverrides)
+
+    // Only the server 0 is available during the bootstrap.
+    val recordMetadata0 = producer.send(new ProducerRecord(topic, part, 0L, "key 0".getBytes, "value 0".getBytes)).get(15, TimeUnit.SECONDS)
+    assertEquals(0, recordMetadata0.offset())
+    adminClient.listTopics().names().get(15, TimeUnit.SECONDS)
+    consumer.assign(Collections.singleton(tp))
+    consumeAndVerifyRecords(consumer, 1, 0)
+
+    server0.shutdown()
+    server0.awaitShutdown()
+    server1.startup()
+
+    assertThrows(classOf[TimeoutException], () => producer.send(new ProducerRecord(topic, part, "key 2".getBytes, "value 2".getBytes)).get(5, TimeUnit.SECONDS))
+    assertThrows(classOf[TimeoutException], () => adminClient.listTopics().names().get(5, TimeUnit.SECONDS))
+
+    val producer2 = createProducer(configOverrides = configOverrides)
+    producer2.send(new ProducerRecord(topic, part, 1L, "key 1".getBytes, "value 1".getBytes)).get(15, TimeUnit.SECONDS)
+    assertEquals(0, consumer.poll(Duration.ofSeconds(5)).count)
+  }
+
   private def sendRecords(numRecords: Int, from: Int): Unit = {
     val producer: KafkaProducer[Array[Byte], Array[Byte]] = createProducer()
     (from until (numRecords + from)).foreach { i =>
@@ -91,6 +130,8 @@ class ConsumerRebootstrapTest extends RebootstrapTest {
 }
 
 object ConsumerRebootstrapTest {
+
+  final val RebootstrapTestName = s"${TestInfoUtils.TestWithParameterizedQuorumAndGroupProtocolNames}.useRebootstrapTriggerMs={2}"
   def rebootstrapTestParams: stream.Stream[Arguments] = {
     assertEquals(1, getTestQuorumAndGroupProtocolParametersClassicGroupProtocolOnly_ZK_implicit.count())
     val args = getTestQuorumAndGroupProtocolParametersClassicGroupProtocolOnly_ZK_implicit
@@ -101,4 +142,3 @@ object ConsumerRebootstrapTest {
     )
   }
 }
-
